% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multireport.R
\name{multireport}
\alias{multireport}
\title{Generate multiple reports using parameterized reporting}
\usage{
multireport(
  rmarkdown_file,
  param_all_options,
  param_name,
  report_format = "html_document",
  report_suffix = "report",
  report_output_dir = "inst"
)
}
\arguments{
\item{rmarkdown_file}{The RMarkdown file to be rendered}

\item{param_all_options}{All possible options for the parameter (e.g. all 50 states)}

\item{param_name}{The name of the parameter to be used for parameterized reporting}

\item{report_format}{The format of the report (html_document, pdf_document, word_document, etc)}

\item{report_suffix}{The name of the prefix of the report names}

\item{report_output_dir}{The name of the directory where the reports will be rendered to}
}
\value{

}
\description{
A function to simply parameterized reporting
}
