% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multireport.R
\name{multireport}
\alias{multireport}
\title{Title}
\usage{
multireport(
  rmarkdown_file,
  params_all_options,
  param_name,
  report_format = "html_document",
  report_suffix = "report",
  report_output_dir = "inst"
)
}
\arguments{
\item{report_suffix}{}
}
\value{

}
\description{
Title
}
