% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multireport.R
\name{multireport}
\alias{multireport}
\title{Generate multiple reports using parameterized reporting}
\usage{
multireport(
  rmarkdown_file,
  params_data_frame,
  report_format = "html_document",
  report_prefix = "",
  report_suffix = "",
  report_output_directory = here::here()
)
}
\arguments{
\item{rmarkdown_file}{The RMarkdown file to be rendered}

\item{params_data_frame}{A data frame with all of the parameter options (each variable name is the name of a parameter and all of the rows in that column are the parameter options)}

\item{report_format}{The format of the report (html_document, pdf_document, word_document, etc)}

\item{report_prefix}{Text to be put before the file names of rendered reports}

\item{report_suffix}{Text to be put fterfile names of rendered reports}

\item{report_output_dir}{The name of the directory where the reports will be rendered to}
}
\value{

}
\description{
A function to simply parameterized reporting
}
